Adapted Stun and Speedy Attack
Changed return type of Stun and Speedy Attack Methods from void to bool to flag if they were sucessful or not in BattleManager. This makes it possible to alter Battle Commentary and display relevant text based on if these actions were sucessful or not.

Also fixed a bug where player action menu stayed active while waiting for turn to change which would allow the player to spam attacks.