Pass Enemy Stats From Overworld To Battle
Player and Game Manager store references to eachother.
Player reads enemy stats when they collide in overowrld, then sends the stats to the game manager by calling GameManager.LoadBattle and sending the stats (stored in a dictionary) as a parameter.

GameManager stores the stats and loads the battle Scene.

BattleManager now seaches for GameManager and stores a reference to it. The calls method to retieve and assign these stats to the enemy during setup.

This allows for one battle scene and one lot of code to be resued for any enemy with any stats.