Finalised NPC behaviour
NPC's now have their own individual UI Canvas to activate and display dialogue on. This ensures in the most simple way that references to the UI are not lost between scene loads and also allows scope for unique UI designs for different types of NPC, rather than one generic dialogue box for all if wanted.