Altered implementation of begining battle
Battle now initiated by player colliding with box collider (which is a trigger) assigned to enemies. This is due to the previous implementation (colliding with the enemy physically) did not work with the final enemy model due to the sword and shield colliders getting in the way. Using this trigger collider means that battle is initiated when the player is one or two steps away from the enemy which is better anyway.