Altered how enemy stats are processed and passed around
Enemy stats are now accessed and stored solely in GameManager. previously, they were assigned to a dictionary in PlayerController which was then passed to GameManager and assigned to a second dictionary. This was inefficient. Now, the enemy the player collides with is passed to the GameManager and the stats are accessed here.

This sets up for the next task which is to deactivate an enemy in the overworld if it is defeated in battle.