Remove defeated enemies from overworld
Enemies that have been fought by the player and defeated are now not active in the overoworld upon returning. This is achieved by storing a list of all enemies and a list of the ID's of defeated enemies. When the overworld is loaded the lists are compared and the enemies that have been defeated are ste to be inactive. Also implemented is the Unity scene loading event system. This solved a large problem where by the operation above was being performed before the overworld was fully loaded and thus not working properly. The unity scene mabger contains a system for flagging when scenes are fully loaded to help mitigate issues such as this.