UI Improvements and Collectable Gold Added
UI Buttons and panels designed using Paint.net created and added in to game leading to heavily redesigned UI. Also imported and implemented assets for gold to act as a currency. Gold is collectable and quanitity is stored and displayed to player.

Overworld UI tied to player and is hidden when in battle.

Bug to fix: gold reloaded when player returns to overworld - should remain invisible if collected.