Consumable Item Implemented
Check is carried out to see if the selected slot contains an instance of ConsumableItem. If it does, as "Consume" button is activated. When clicked, this button checks again to see if the item is aninstance of a consumable item, before casting the item in that slot as a ConsumableItem and storing it in a variable of type ConsumableItem. This is necessary because InventoryItem does not contain a consumeItem() method, and the dictionary is set up to store inventoryItems, so in order to call consumeItem() the reference has to be of type consumableItem.