Fixed Inventory Bug
Fixed inventory bug where objects in inventory were being destroyed. Had to use dont destory on load but tis wouldnt work as the collectable items are children of objects so found a thread on stack overflow that suggested to remove the parent of the object before setting dont destory on load. I am only setting odnt destroy on load on items actually added to the inventory, as all the un-collected ones in the overworld can be destroyed when the scene is unloaded.